This morning: ps 111-116 -sr stephen
Ps 111
OT2—should this be a “Cf”? Verse 13 is about salt; verse 14 is about light. But the French has it reversed.

Ps 112
OT—Is this Jn 14:13 or Jn 14:14?
Chr—add the name Jesus?

Ps 115
OT—should “he” be “Jesus”?	

Ps 116
OT/Adv—if I put in the correct verse, perhaps this should have a “Cf”?
Lent—awkward?